fix acesso indevido

+++ concluir_solicitacao.php
@@ -7,30 +7,37 @@
     unset($_SESSION['privilegio']);
     $_SESSION['msgLogin'] = "<div class='alert alert-danger alert-dismissible fade show' role='alert'>Acesso restrito!<button type='button' class='btn-close' data-bs-dismiss='alert' aria-label='Close'></button></div>"; 
     header('Location: index.php');
-  }
-  $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
-  if (isset($id)) {
-    $confirmaId = true;
   }else {
-    $confirmaId = false;
-  }
-  if ($confirmaId==true) {
-    $select_documentos = "SELECT * FROM `solicitacoes` WHERE id='$id'";
-    $documentos = mysqli_query($connect, $select_documentos);
-    $doc = mysqli_fetch_assoc($documentos);
-    $destino = "documentos/".$doc['documentos'];
-    if(file_exists($destino)) {
-      unlink($destino);
+    if(isset($_SESSION['privilegio']) == true and $_SESSION['privilegio'] != 'Administrador'){
+      $_SESSION['msgLogin'] = "<div class='alert alert-danger alert-dismissible fade show' role='alert'>Acesso somente para Administradores!<button type='button' class='btn-close' data-bs-dismiss='alert' aria-label='Close'></button></div>"; 
+      header("Location: index.php");
+    }else {
+      $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
+      if (isset($id)) {
+        $confirmaId = true;
+      }else {
+        $confirmaId = false;
+      }
+      if ($confirmaId==true) {
+        $select_documentos = "SELECT * FROM `solicitacoes` WHERE id='$id'";
+        $documentos = mysqli_query($connect, $select_documentos);
+        $doc = mysqli_fetch_assoc($documentos);
+        $destino = "documentos/".$doc['documentos'];
+        if(file_exists($destino)) {
+          unlink($destino);
+        }
+        $get_solicitacao = "SELECT * FROM solicitacoes WHERE id=$id";
+        $sol = mysqli_query($connect, $get_solicitacao);
+        $row = mysqli_fetch_assoc($sol);  
+        $solicitacao_concluida = "INSERT INTO solicitacoes_concluidas(id, tipo_certificado, nome, data_solicitacao, contador, data_conclusao) VALUES ({$row['id']}, '{$row['tipo_certificado']}', '{$row['nome']}', '{$row['data_solicitacao']}', '{$row['contador']}', NOW())";
+        $solicitacao = mysqli_query($connect, $solicitacao_concluida);
+        $apagar_solicitacao = "DELETE FROM solicitacoes WHERE id='$id'";
+        $apagar = mysqli_query($connect, $apagar_solicitacao);
+        $_SESSION['concluirSolicitacao'] = "<div class='alert alert-success alert-dismissible fade show' role='alert'>Solicitação concluida com sucesso!<button type='button' class='btn-close' data-bs-dismiss='alert' aria-label='Close'></button></div>";    
+        header("Location: solicitacoes_ativas.php");  
+        $confirmaId = false;
+      }
     }
-    $get_solicitacao = "SELECT * FROM solicitacoes WHERE id=$id";
-    $sol = mysqli_query($connect, $get_solicitacao);
-    $row = mysqli_fetch_assoc($sol);  
-    $solicitacao_concluida = "INSERT INTO solicitacoes_concluidas(id, tipo_certificado, nome, data_solicitacao, contador, data_conclusao) VALUES ({$row['id']}, '{$row['tipo_certificado']}', '{$row['nome']}', '{$row['data_solicitacao']}', '{$row['contador']}', NOW())";
-    $solicitacao = mysqli_query($connect, $solicitacao_concluida);
-    $apagar_solicitacao = "DELETE FROM solicitacoes WHERE id='$id'";
-	  $apagar = mysqli_query($connect, $apagar_solicitacao);
-    $_SESSION['concluirSolicitacao'] = "<div class='alert alert-success alert-dismissible fade show' role='alert'>Solicitação concluida com sucesso!<button type='button' class='btn-close' data-bs-dismiss='alert' aria-label='Close'></button></div>";    
-    header("Location: solicitacoes_ativas.php");  
-    $confirmaId = false;
+    header("Location: solicitacoes_concluidas.php");
   }
 ?>
