+++ excluir_usuario.php
@@ -7,9 +7,15 @@
     unset($_SESSION['privilegio']);
     $_SESSION['msgLogin'] = "<div class='alert alert-danger alert-dismissible fade show' role='alert'>Acesso restrito!<button type='button' class='btn-close' data-bs-dismiss='alert' aria-label='Close'></button></div>"; 
     header('Location: index.php');
+  }else {
+    if(isset($_SESSION['privilegio']) == true and $_SESSION['privilegio'] != 'Administrador'){
+      $_SESSION['msgLogin'] = "<div class='alert alert-danger alert-dismissible fade show' role='alert'>Acesso somente para Administradores!<button type='button' class='btn-close' data-bs-dismiss='alert' aria-label='Close'></button></div>"; 
+      header("Location: index.php");
+    }else {
+      $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
+      $excluir_usuario = "DELETE FROM usuarios WHERE id='$id'";
+      $excluir = mysqli_query($connect, $excluir_usuario);   
+      header("Location: usuarios.php");    
+    }
   }
-  $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
-  $excluir_usuario = "DELETE FROM usuarios WHERE id='$id'";
-  $excluir = mysqli_query($connect, $excluir_usuario);   
-  header("Location: usuarios.php");    
 ?>
